pyproject.toml
[project]
name = "sistemainventariokasu"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.42.93",
    "dj-database-url>=3.1.2",
    "django==5.2",
    "django-apscheduler>=0.7.0",
    "django-storages>=1.14.6",
    "gunicorn>=25.3.0",
    "pillow>=12.2.0",
    "psycopg2-binary>=2.9.12",
    "python-dateutil>=2.9.0.post0",
    "python-decouple>=3.8",
    "qrcode[pil]>=8.2",
    "sendgrid>=6.12.5",
    "weasyprint>=68.1",
    "whitenoise>=6.12.0",
]
